Parse GA4GH Retrieval protocol and handle redirects
Accessing a file via the GA4GH Retrieval protocol returns a JSON ticket
containing a list of URLs whose contents are to be concatenated to form
the file to be retrieved.

Add a multipart hFILE pseudo-backend ("pseudo-" in that it is not
dispatched via a scheme lookup in hopen() itself) that handles reading
from the concatenation of a list of files.

Parse the GA4GH JSON text to initialise this special type of hFILE.

In hts_hopen(), if JSON format is detected, reopen by "redirecting" to
the multipart hFILE specified in the GA4GH ticket file.  With two hFILE
handles in play, care must be taken to close the multipart hFILE if
hts_hopen() subsequently fails, and to close the ticket hFILE only if
hts_hopen() succeeds.  Also ensure all error paths set errno.