Fix hFILE write-after-read bug
When an	  hFILE is open read-write, ensure that writes that follow reads go
to the correct location.  This fixes bugs where read, seek backward,
write would either not be flushed out, or would	flush to the wrong location.

When a write occurs that starts	  before the current end pointer (so
previously read	    data is	  being overwritten):
* Make to backend file pointer seek to where the write will start.
* Discard anything currently in the buffer.