Make libcurl_seek work better when servers can't do byte ranges
Duplicate the curl easy handle so that a new connection is opened
when attempting a request on a byte range.  If the new connection
works, the old one is closed and replaced by the new one.  If the
new connection fails (for example because the data is coming from
a CGI script that can't honour byte ranges), close it and
continue with the original connection.  In the latter case errno
is set to ESPIPE so the caller can fall back to whatever it does
when working on a stream.

Writing to byte ranges is ill-defined for HTTP, so we disable seek
on handles opened for write by returning -1 and setting errno to
ESPIPE in all cases.  Arguable it might work for other protocols
(e.g. FTP) but we probably don't attempt it at the moment.

Fixes problem where the BAM EOF check would cause loss of
connection to servers that didn't honour the Range: HTTP header.