Make the sam parser detect out-of-range numbers
Replace STRTOL64() and STRTOUL64() functions with versions that can
detect when the parsed value will not fit into the destination.
The new functions are called hts_str2int() and hts_str2uint() as the
interface has changed somewhat from that of strtol().  The new
functions have been put in textutils_internal.h so that they
can be used by other parsers (notably vcf) in future updates.

Add tests for the new hts_str2int() and hts_str2uint() functions.

Update sam_parse1() to use these functions and report an error if
a number was too large.  The B array code is pulled out into its
own function.  The new function includes code to fix up B arrays
where the subtype code was set incorrectly (at the cost of a couple
of extra scans through the data).

Put a struct hFILE declaration in textutils_internal.h so that
it can be included on its own.