Fix illegal shifts in ltf8_decode_crc; cull legacy functions
Fix cases of shift into sign bit in ltf8_decode_crc().  For 8
and 9 byte codings, it is not necessary to include any of the
bits from the first byte read as they will either be always 0
(for 8 byte) or are not used at all (9 byte).  The 9 byte
coding is also altered to use an unsigned accumulator to avoid
the shift into sign bit problem (as was done earlier in
itf8_decode_crc).

Remove legacy ltf8_decode() and itf8_decode() functions that
ar no longer used.

Fix double semicolons in ltf8_decode_crc()