Fix possible integer wrap-around in bam_cigar2rqlens
Convert it to use uint64_t * for the returned lengths.  This
prevents the accumulated length from wrapping around.

bam_cigar2rqlens is not yet part of the API, so the change does
not affect the public ABI.

Move the check on invalid l_qseq in bam_read1() up slightly to
reduce unnecessary work if the test fails.