Prevent possible buffer underflow in rans_uncompress_O1()
When calculating the offset to an output stream, the value
`3*isz4` could wrap to a negative value.  This is then used
as an index to `out_buf`, allowing arbitrary data to be written
to a location before the start of the output buffer.  Fix
by using uint32_t instead of int.

Make some integers in rans_uncompress_O0() unsigned as well
to avoid possible, albeit less serious, wrap-around issues.

Make RansDecRenormSafe() return an error code if it has run out
of input data, and make the uncompress functions bail out
if this happens.  This stops the uncompressor from spending
ages in a complicated and useless memset if `out_sz` is much
bigger than the available data can supply.

Bug found using libfuzzer.