Ensure s->s is allocated in kputsn_ even if l is zero
Make sure that ks_resize() is called with a value greater than
zero so that a buffer is always allocated if none was present
before.  Prevents passing a NULL pointer to memcpy, which is
undefined behaviour.  Also prevents possible undefined behaviour
in callers that expect something to have been allocated and so
don't bother to check if s->s is NULL.