Use malloc instead of calloc for targets array
While it usually doesn't make any difference, under certain
conditions glibc uses sbrk() and memset for calloc which can
be slow for large arrays.

This is safe as a count is kept of the number of entries that
have been allocated so the correct number can be freed if an error
occurs.  When the function returns either all the target_name
entries will contain a valid pointer, or everything will have
been cleaned up.