Handle q==0.0 case in kt_fisher_exact() and add tests
When some of kt_fisher_exact()'s input arguments are very large,
hypergeo(...) is dominated by the subtracted lbinom() so returns
exp(-750 or less), which underflows, hence hypergeo() returns 0.0.

When q is 0.0, kt_fisher_exact()'s loops never execute and which of i
and j is closer to n11 is indeterminate. To avoid this problem a
special case is added for q == 0.0 which determines the output
values by comparing n11 to the location of the peak of the
hypergeometric distribution.

Co-Authored-By: Rob Davies <rmd+git@sanger.ac.uk>