Improve windows build.
Specifically we create the extra files needed for MSVC linkage, and
document the MSYS2/MINGW setup process.

Also added a win-dist target which attempts to produce a directory
structure suitable for binary distribution.  This isn't executed by
default, but is a good aide-memoire and to simplify testing
compatibility with things like MSVC.

Ideally we'd have a similar mechanism for all platforms to permit easy
creation of binary distributions (see #533).