Fix unary-not on null strings in expressions
Toggling hts_expr_val::is_true on strings could get it out of
phase with hts_expr_val::d on null strings (which are false),
which caused double-unary-not to give the wrong value.

Instead, make unary not always return false if is_true is true,
so empty-but-true works; and for strings return true for null
ones, and false for non-null.  Numbers are handled as before.