Detect uninitialised and recycled pointers in hts_filter_eval()
Due to hts_filter_eval() calling memset() on its res parameter,
it's not possible to pass in an allocted kstring_t in res->s
without leaking memory.  Historically it was also possible
to get away with passing in an uninitialised structure, so not
many assumptions can be made about the contents of res on entry.
In particular, it is not guaranteed that free(res->s.s) would
work.

To ensure the function is being used safely, check that the
string part of *res is NULL on entry and fail ungracefully if not.
If it does fail, the code calling it needs to be rewritten, so
it seems reasonable to call abort here.  (It also seems that
little if anything outside of HTSlib uses the interface directly
anyway, so hopefully we won't break any external code with
this change).  Code passing the tests will work correctly even if
linked against earlier versions of HTSlib.