Fix bug where bin number could overflow when looking for max_off
When searching for `max_off`, hts_itr_query() and
hts_itr_multi_bam() look for a bin to the right of the end of
the region.  For whole chromosomes, this would be HTS_POS_MAX,
which is far beyond the maximum bin position supported.  The `bin`
calculation overflowed leading to either a negative bin number
or an incorrect positive one, depending on the number of levels
in the index.  Negative bin numbers simply caused time to be
wasted as the search loop eventually counted up to zero, but
incorrect positive ones could cause the iterator to finish too
early.

Fix by catching the out-of-bounds case and setting max_off
to UINT64_MAX, whch should be used for bins beyond the end
of the indexable range.